Delete NetworkMonitor in favour of Reachability. Make Reachability Observable
