+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Tabs/TabTray/TabTrayController.swift
@@ -919,7 +919,7 @@ class TabTrayController: AuthenticationController {
         )
       )
     case .openTabsDisabled, .noSyncedSessions:
-      if !DeviceInfo.hasConnectivity {
+      if Reachability.shared.isOffline {
         present(SyncAlerts.noConnection, animated: true)
         return
       }
