+++ ios/brave-ios/Sources/Brave/Frontend/Settings/SettingsViewController.swift
@@ -473,7 +473,7 @@ class SettingsViewController: TableViewController {
           text: Strings.Sync.syncTitle,
           selection: { [unowned self] in
             if syncAPI.isInSyncGroup {
-              if !DeviceInfo.hasConnectivity {
+              if Reachability.shared.isOffline {
                 self.present(SyncAlerts.noConnection, animated: true)
                 return
               }
