+++ ios/brave-ios/Sources/Brave/Frontend/Sync/SyncViewController.swift
@@ -71,7 +71,7 @@ class SyncViewController: AuthenticationController {
   /// Perform a block of code only if user has a network connection, shows an error alert otherwise.
   /// Most of sync initialization methods require an internet connection.
   func doIfConnected(code: () -> Void) {
-    if !DeviceInfo.hasConnectivity {
+    if Reachability.shared.isOffline {
       present(SyncAlerts.noConnection, animated: true)
       return
     }
