+++ ios/brave-ios/Sources/BraveVPN/Components/Settings/BraveVPNContactFormViewController.swift
@@ -3,6 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+import BraveShared
 import BraveStrings
 import BraveUI
 // To get cellular carrier name
